% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_añadir_percentil.R
\name{adi.perc}
\alias{adi.perc}
\title{Añadir Percentil}
\usage{
adi.perc(x, y = 0)
}
\arguments{
\item{x}{Data Frame de los datos de referencia}

\item{y}{Data Frame de nuevos datos}
}
\value{
Si se introduce una base de datos, un data frame con los datos
originales y los datos de percentiles. Si se introducen dos bases de datos,
una lista con dos data frames, correspondientes a los dos introducidos.
}
\description{
A partir de una base de datos, por cada columna con datos numéricos, se
añadirá una nueva columna con los valores de percentiles de la columna
original. Si se añade una segunda base de datos se realizará la adición de
los datos de percentiles a ambas bases de datos, obteniéndose los valores a
asignar a partir de la primera base de datos.
}
