% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_calcular_percentil.R
\name{calc.perc}
\alias{calc.perc}
\title{Calcular percentil}
\usage{
calc.perc(x)
}
\arguments{
\item{x}{Vector numérico}
}
\value{
Vector numérico con los respectivos percentiles
}
\description{
A partir de una vector se calculan los umbrales de cada percentil para luego
crear un nuevo vector con los valores percentiles de cada valor original
}
