% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod1_cargar_datos.R
\name{cargar_datosServer}
\alias{cargar_datosServer}
\title{Módulo 1: Carga de datos (Server)}
\usage{
cargar_datosServer(id)
}
\arguments{
\item{id}{Identificador}
}
\description{
Módulo 1: Carga de datos (Server)
}
