% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod1_cargar_datos.R
\name{cargar_datosUI}
\alias{cargar_datosUI}
\title{Módulo 1: Carga de datos (UI)}
\usage{
cargar_datosUI(id)
}
\arguments{
\item{id}{Identificador}
}
\description{
Permite realizar la carga de datos de referencia y los datos a testear y
visualizarlos en sus respectivos marcos de datos.
}
