% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{FootManScoutApp}
\alias{FootManScoutApp}
\title{App para el análisis de datos estadísticos con finalidades de Scouting}
\usage{
FootManScoutApp()
}
\arguments{
\item{input}{Input}

\item{output}{Output}
}
\description{
Permite realizar diferentes análisis de las estadísticas recopiladas durante
la temporada para facilitar la búsqueda de jugadores. La aplicación se divide
en módulos, cada uno de los cuales se centra en un análisis diferente. El
Módulo 1 se encarga de la carga de los datos; el Módulo 2 se encarga del
análisis de percentiles; el Módulo 3 se encarga del análisis gráfico.
}
